Try to read directory

+++ P1/src/structs/ent_ar.c
@@ -1,7 +1,7 @@
 #include "ent_ar.h"
 #include <stdlib.h>
 
-EntAr* entar_init(char validez, unsigned long int identificador_relativo, unsigned long int identificador_absoluto, char* nombre_archivo) {
+EntAr* entar_init(char validez, signed int identificador_relativo, signed int identificador_absoluto, char* nombre_archivo) {
     EntAr* entar = malloc(sizeof(EntAr));
     entar -> validez = (int) validez;
     entar -> identificador_relativo = identificador_relativo;
