+++ P1/src/structs/ent_ar.h
@@ -5,15 +5,15 @@
 typedef struct ent_ar
 {
     int validez; // se castea desde un char
-    unsigned long int identificador_relativo;
-    unsigned long int identificador_absoluto;
+    signed int identificador_relativo;
+    signed int identificador_absoluto;
     Indice* indice;
     char* nombre_archivo;
 } EntAr;
 
 // FUNCTIONS
 
-EntAr* entar_init(char validez, unsigned long int identificador_relativo, unsigned long int identificador_absoluto, char* nombre_archivo);
+EntAr* entar_init(char validez, signed int identificador_relativo, signed int identificador_absoluto, char* nombre_archivo);
 
 void assign_indice(EntAr* entar, Indice* indice);
