+++ P1/src/structs/ent_dir.h
@@ -4,16 +4,16 @@
 typedef struct ent_dir {
     int validez; // es un bit 0 o 1
     int identificador_particion;
-    unsigned long int identificador_directorio;
-    unsigned long int cantidad_bloques_particion;
-    unsigned long int cantidad_bitmaps;
+    signed int identificador_directorio;
+    signed int cantidad_bloques_particion;
+    signed int cantidad_bitmaps;
     BitMap** lista_de_bitmaps;
 } EntDir;
 
 EntDir* entdir_init(char validez,
                     int identificador_particion,
-                    unsigned long int identificador_directorio,
-                    unsigned long int cantidad_bloques_particion);
+                    signed int identificador_directorio,
+                    signed int cantidad_bloques_particion);
 
 
 void entdir_clean(EntDir* entdir);
