+++ P1/src/structs/indice.h
@@ -2,17 +2,17 @@
 #include "data.h"
 
 typedef struct indice {
-    unsigned long int tamano; // tamano archivo
+    signed int tamano; // tamano archivo
     int cantidad_bloques; // tamaño dividido por 2kb
     Data** lista_de_datos; // lista con numeros de bloques
-    unsigned long int identificador_relativo;
-    unsigned long int identificador_absoluto;
+    signed int identificador_relativo;
+    signed int identificador_absoluto;
 } Indice;
 
 
 // FUNCTIONS
 
-Indice* indice_init(unsigned long int tamano, unsigned long int identificador_relativo, unsigned long int identificador_absoluto);
+Indice* indice_init(signed int tamano, signed int identificador_relativo, signed int identificador_absoluto);
 
 void assing_data_list(Indice* indice, Data* data, int index);
